+++ src/main/resources/db/changelog/changes/v0001_Create_Table_Classificacao.sql
@@ -1,11 +1,11 @@
 create table Classificacao (
     id                  serial not null,
     time                varchar (100) not null,
-    gols_feitos          integer,
-    gols_levados         integer,
-    quantidade_over25    integer,
-    quantidade_under25   integer,
-    ambas_marcam   integer,
+    gols_feitos          BIGINT,
+    gols_levados         BIGINT,
+    quantidade_over25    BIGINT,
+    quantidade_under25   BIGINT,
+    ambas_marcam   BIGINT,
     liga                varchar (100),
     primary key (id)
 )
